% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_author.R
\name{add_author}
\alias{add_author}
\title{Add an author to a file}
\usage{
add_author(path = ".", role = c("aut", "rev", "cph", "fnd"))
}
\arguments{
\item{path}{A path to a file or directory.
The file must be a quarto file (e.g. \verb{_quarto.yml} or \verb{*.qmd}).}

\item{role}{The role of the person to add.
One of \code{"aut"} (author), \code{"rev"} (reviewer), \code{"cph"} (copyright holder), or
\code{"fnd"} (funder).}
}
\description{
Add an author to a file
}
