% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_individual.R
\name{add_individual}
\alias{add_individual}
\title{Add an individual to a file}
\usage{
add_individual(
  path = ".",
  role = c("aut", "cre", "ctb", "rev", "cph", "fnd", "pbl"),
  org = org_list$new()$read()
)
}
\arguments{
\item{path}{A path to a file or directory.
Supported file types:
\itemize{
\item Quarto files (\verb{_quarto.yml} or \verb{*.qmd})
\item R package DESCRIPTION files
\item README files (\code{README.md} or \code{README.Rmd})
\item Bookdown index files (\code{index.md} or \code{index.Rmd})
}}

\item{role}{The role of the person to add.
One or multiple of \code{"aut"} (author), \code{"cre"} (creator/maintainer), \code{"ctb"}
(contributor), \code{"rev"} (reviewer), \code{"cph"} (copyright holder),
\code{"fnd"} (funder), or \code{"pbl"} (publisher).
Unless you want to an individual to a Quarto or bookdown document.
Then you can only specify one role, and \code{"aut"} will be added to the \code{author}
field, #' \code{"rev"} to the \code{reviewer} field, \code{"cph"} to the \code{rightsholder}
field, \code{"fnd"} to the \code{funder} field, and \code{"pbl"} to the \code{publisher} field.}

\item{org}{An \code{org_list} object containing the organisation information.
Defaults to the \code{org_list} object read from the current working directory.}
}
\description{
Add an individual to a file
}
\examples{
\dontrun{
add_individual(path = ".", role = "aut")
}
}
\seealso{
Other individual:
\code{\link[=has_person_role]{has_person_role()}},
\code{\link[=individual2badge]{individual2badge()}},
\code{\link[=individual2df]{individual2df()}},
\code{\link[=individual2person]{individual2person()}},
\code{\link[=select_individual]{select_individual()}},
\code{\link[=select_person_role]{select_person_role()}},
\code{\link[=store_individuals]{store_individuals()}}
}
\concept{individual}
