% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ask_keywords.R
\name{ask_keywords}
\alias{ask_keywords}
\title{Ask for keywords
This function prompts the user to enter one or more keywords separated by
\verb{;}.
It will continue to prompt the user until at least one keyword is provided.
The function returns a character vector of the entered keywords, with leading
and trailing whitespace removed.}
\usage{
ask_keywords()
}
\value{
A character vector of keywords entered by the user.
}
\description{
Ask for keywords
This function prompts the user to enter one or more keywords separated by
\verb{;}.
It will continue to prompt the user until at least one keyword is provided.
The function returns a character vector of the entered keywords, with leading
and trailing whitespace removed.
}
\examples{
\dontrun{
keywords <- ask_keywords()
}
}
\seealso{
Other question:
\code{\link[=ask_email]{ask_email()}},
\code{\link[=ask_language]{ask_language()}},
\code{\link[=ask_new_license]{ask_new_license()}},
\code{\link[=ask_orcid]{ask_orcid()}},
\code{\link[=ask_ror]{ask_ror()}},
\code{\link[=ask_url]{ask_url()}},
\code{\link[=ask_yes_no]{ask_yes_no()}},
\code{\link[=menu_first]{menu_first()}}
}
\concept{question}
