% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ask_language.R
\name{ask_language}
\alias{ask_language}
\title{Ask for a language}
\usage{
ask_language(languages, prompt = "Which language?")
}
\arguments{
\item{languages}{A character vector containing the available languages.}

\item{prompt}{the string printed when prompting the user for input.
  Should usually end with a space \code{" "}.}
}
\value{
A character string containing a valid language code.
}
\description{
Ask for a language
}
\examples{
\dontrun{
ask_language(languages = c("en-GB", "nl-BE"))
}
}
\seealso{
Other question:
\code{\link[=ask_email]{ask_email()}},
\code{\link[=ask_keywords]{ask_keywords()}},
\code{\link[=ask_new_license]{ask_new_license()}},
\code{\link[=ask_orcid]{ask_orcid()}},
\code{\link[=ask_ror]{ask_ror()}},
\code{\link[=ask_url]{ask_url()}},
\code{\link[=ask_yes_no]{ask_yes_no()}},
\code{\link[=menu_first]{menu_first()}}
}
\concept{question}
