% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ask_ror.R
\name{ask_ror}
\alias{ask_ror}
\title{Ask for an ROR}
\usage{
ask_ror(prompt)
}
\arguments{
\item{prompt}{the string printed when prompting the user for input.
  Should usually end with a space \code{" "}.}
}
\value{
A character string containing a valid ROR identifier or an empty
string.
}
\description{
The Research Organization Registry (ROR) is a global, community-led registry
of open persistent identifiers for research organizations.
ROR makes it easy for anyone or any system to disambiguate institution names
and connect research organizations to researchers and research outputs.
}
\examples{
\dontrun{
ask_ror("Enter ROR: ")
}
}
\seealso{
Other question:
\code{\link[=ask_email]{ask_email()}},
\code{\link[=ask_keywords]{ask_keywords()}},
\code{\link[=ask_language]{ask_language()}},
\code{\link[=ask_new_license]{ask_new_license()}},
\code{\link[=ask_orcid]{ask_orcid()}},
\code{\link[=ask_url]{ask_url()}},
\code{\link[=ask_yes_no]{ask_yes_no()}},
\code{\link[=menu_first]{menu_first()}}
}
\concept{question}
