% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ask_url.R
\name{ask_url}
\alias{ask_url}
\title{Ask an URL}
\usage{
ask_url(prompt)
}
\arguments{
\item{prompt}{the string printed when prompting the user for input.
  Should usually end with a space \code{" "}.}
}
\value{
A character string containing a valid URL or an empty string.
}
\description{
Ask an URL
}
\examples{
\dontrun{
ask_url("Enter URL: ")
}
}
\seealso{
Other question:
\code{\link[=ask_email]{ask_email()}},
\code{\link[=ask_keywords]{ask_keywords()}},
\code{\link[=ask_language]{ask_language()}},
\code{\link[=ask_new_license]{ask_new_license()}},
\code{\link[=ask_orcid]{ask_orcid()}},
\code{\link[=ask_ror]{ask_ror()}},
\code{\link[=ask_yes_no]{ask_yes_no()}},
\code{\link[=menu_first]{menu_first()}}
}
\concept{question}
