% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_org.R
\name{cache_org}
\alias{cache_org}
\title{Cache organisation list from URL}
\usage{
cache_org(url, config_folder = R_user_dir("citeme", "config"))
}
\arguments{
\item{url}{The URL of the organisation list to retrieve.}

\item{config_folder}{The folder where the cached organisation list should be
stored.
Defaults to the user's R configuration directory for \code{citeme}.}
}
\value{
The retrieved organisation list, or \code{NULL} if the URL is invalid or
the repository.
}
\description{
This function retrieves the organisation list from a given URL and caches it
locally.
If the URL is \code{"https://github.com/inbo"}, it uses the default INBO
organisation list.
For other URLs, it checks if a public \code{citeme} repository exists and clones
it to retrieve the organisation list.
The cached organisation list is stored in a specified configuration folder.
}
\examples{
\dontrun{
cache_org("https://github.com/inbo")
}
}
\seealso{
Other organisation:
\code{\link[=get_available_organisations]{get_available_organisations()}},
\code{\link[=get_default_org_list]{get_default_org_list()}},
\code{\link[=inbo_org_list]{inbo_org_list()}},
\code{\link[=new_org_item]{new_org_item()}},
\code{\link[=new_org_list]{new_org_list()}},
\code{\link[=org_list_from_url]{org_list_from_url()}},
\code{\link[=organisation2df]{organisation2df()}},
\code{\link[=select_license]{select_license()}},
\code{\link[=store_organisations]{store_organisations()}},
\code{\link[=stored_organisations]{stored_organisations()}}
}
\concept{organisation}
