% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coalesce.R
\name{coalesce}
\alias{coalesce}
\title{coalesce}
\usage{
coalesce(...)
}
\arguments{
\item{...}{Arguments to check for non-NULL values.}
}
\value{
The first non-NULL argument, or NULL if all arguments are NULL.
}
\description{
Return the first non-NULL argument.
}
\examples{
coalesce(NULL, NULL, "a", "b")
coalesce(NULL, 1)
coalesce(NULL)
}
\seealso{
Other utils:
\code{\link[=add_badges]{add_badges()}}
}
\concept{utils}
