% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_default_org_list.R
\name{get_default_org_list}
\alias{get_default_org_list}
\title{Get the default organization list}
\usage{
get_default_org_list(x = ".")
}
\arguments{
\item{x}{The path to the repository.
Defaults to the current working directory.}
}
\value{
An \code{org_list} object containing the organisation list.
The function also stores the information in the user's R configuration.
}
\description{
This function retrieves the default organisation list from the
\code{organisation.yml} file in the organisations \code{citeme} repository.
The \code{origin} of the repository is used to determine the root URL of the
organisation.
}
\examples{
\dontrun{
get_default_org_list()
}
}
\seealso{
Other organisation:
\code{\link[=cache_org]{cache_org()}},
\code{\link[=get_available_organisations]{get_available_organisations()}},
\code{\link[=inbo_org_list]{inbo_org_list()}},
\code{\link[=new_org_item]{new_org_item()}},
\code{\link[=new_org_list]{new_org_list()}},
\code{\link[=org_list_from_url]{org_list_from_url()}},
\code{\link[=organisation2df]{organisation2df()}},
\code{\link[=select_license]{select_license()}},
\code{\link[=store_organisations]{store_organisations()}},
\code{\link[=stored_organisations]{stored_organisations()}}
}
\concept{organisation}
