% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has_person_role.R
\name{has_person_role}
\alias{has_person_role}
\title{Does a person object has a given role?}
\usage{
has_person_role(individual, role)
}
\arguments{
\item{individual}{A \code{person} object.}

\item{role}{A character vector of roles to check for.}
}
\value{
A logical vector indicating whether the individual has any of the
specified roles.
}
\description{
Does a person object has a given role?
}
\examples{
p <- person("Jane", "Doe", role = c("aut", "cre"))
has_person_role(p, "aut")
has_person_role(p, "ctb")
}
\seealso{
Other individual:
\code{\link[=add_individual]{add_individual()}},
\code{\link[=individual2badge]{individual2badge()}},
\code{\link[=individual2df]{individual2df()}},
\code{\link[=individual2person]{individual2person()}},
\code{\link[=select_individual]{select_individual()}},
\code{\link[=select_person_role]{select_person_role()}},
\code{\link[=store_individuals]{store_individuals()}}
}
\concept{individual}
