% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/org_list_class.R
\name{inbo_org_list}
\alias{inbo_org_list}
\title{The INBO organisation list}
\usage{
inbo_org_list()
}
\description{
The INBO organisation list
}
\seealso{
Other organisation:
\code{\link[=cache_org]{cache_org()}},
\code{\link[=get_available_organisations]{get_available_organisations()}},
\code{\link[=get_default_org_list]{get_default_org_list()}},
\code{\link[=new_org_item]{new_org_item()}},
\code{\link[=new_org_list]{new_org_list()}},
\code{\link[=org_list_from_url]{org_list_from_url()}},
\code{\link[=select_license]{select_license()}}
}
\concept{organisation}
