% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individual2badge.R
\name{individual2badge}
\alias{individual2badge}
\title{Create a markdown badge for an individual}
\usage{
individual2badge(
  individual,
  role = c("aut", "cre", "cph", "ctb", "fnd", "pbl", "rev")
)
}
\arguments{
\item{individual}{A data frame with columns \code{given}, \code{family}, \code{orcid},
\code{email}, and \code{affiliation}.}

\item{role}{A character string indicating the individual's role.
Must be one or more of \code{"aut"} (author), \code{"cre"} (contact person),
\code{"cph"} (copyright holder), \code{"ctb"} (contributor), \code{"fnd"} (funder),
\code{"rev"} (reviewer).
Default is `"aut"``.}
}
\value{
A character string containing the markdown badge for the individual.
}
\description{
This function creates a markdown badge for an individual, including their
name, ORCID, email, and affiliation if available.
It also adds a footnote with the individual's role and affiliation.
}
\examples{
individual2badge(
  data.frame(
    given = "Jane", family = "Doe", email = "jane@example.com",
    orcid = "0000-0002-1825-0097", affiliation = "University"
  ),
  role = "aut"
)
}
\seealso{
Other individual:
\code{\link[=add_individual]{add_individual()}},
\code{\link[=has_person_role]{has_person_role()}},
\code{\link[=individual2df]{individual2df()}},
\code{\link[=individual2person]{individual2person()}},
\code{\link[=select_individual]{select_individual()}},
\code{\link[=select_person_role]{select_person_role()}},
\code{\link[=store_individuals]{store_individuals()}}
}
\concept{individual}
