% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individual2person.R
\name{individual2person}
\alias{individual2person}
\title{Select the individual information to use as a \code{person} object}
\usage{
individual2person(individual, role = "aut", lang, org = org_list$new()$read())
}
\arguments{
\item{individual}{An optional \code{individual} object to convert to a \code{person}
object.
If not provided, the function will prompt the user to select an individual.}

\item{role}{The role to use for the \code{person} object.
Defaults to `"aut"`` (author).}

\item{lang}{The language to use for the affiliation.}

\item{org}{An \code{org_list} object containing the organisation information.
Defaults to the \code{org_list} object read from the current working directory.}
}
\value{
A \code{\link[utils:person]{person}} object with the given name, family name,
email, ORCID, and affiliation of the selected individual.
}
\description{
This function retrieves the individual information using
\code{select_individual()} and converts it to a \code{person} object.
If the \code{individual} argument is not provided, it will prompt the user to
select an individual.
The \code{person} object will include the given name, family name, email, ORCID,
and affiliation (if available) of the selected individual.
}
\seealso{
Other individual:
\code{\link[=add_individual]{add_individual()}},
\code{\link[=has_person_role]{has_person_role()}},
\code{\link[=individual2badge]{individual2badge()}},
\code{\link[=individual2df]{individual2df()}},
\code{\link[=select_individual]{select_individual()}},
\code{\link[=select_person_role]{select_person_role()}},
\code{\link[=store_individuals]{store_individuals()}}
}
\concept{individual}
