% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_repository.R
\name{is_repository}
\alias{is_repository}
\title{Determine if a directory is in a git repository}
\usage{
is_repository(path = ".")
}
\arguments{
\item{path}{the location of the git repository, see the "path" section.}
}
\value{
TRUE if directory is in a git repository else FALSE
}
\description{
The path arguments specifies the directory at which to start the search for
a git repository.
If it is not a git repository itself, then its parent directory is consulted,
then the parent's parent, and so on.
}
\examples{
is_repository(tempdir())
}
\seealso{
Other git:
\code{\link[=is_tracked_not_modified]{is_tracked_not_modified()}},
\code{\link[=ssh_http]{ssh_http()}}
}
\concept{git}
