% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store_organisations.R
\name{organisation2df}
\alias{organisation2df}
\title{Convert organisation object to a data.frame}
\usage{
organisation2df(x)
}
\arguments{
\item{x}{An \code{org_list} or \code{org_item} object.}
}
\value{
A data.frame with the organisation information.
}
\description{
Results in a data.frame with the email, default name, ROR, ORCID requirement,
Zenodo community, website, and logo of the organisation.
}
\examples{
\dontrun{
org <- org_list$new()$read()
organisation2df(org)
}
}
\seealso{
Other organisation:
\code{\link[=cache_org]{cache_org()}},
\code{\link[=get_available_organisations]{get_available_organisations()}},
\code{\link[=get_default_org_list]{get_default_org_list()}},
\code{\link[=inbo_org_list]{inbo_org_list()}},
\code{\link[=new_org_item]{new_org_item()}},
\code{\link[=new_org_list]{new_org_list()}},
\code{\link[=org_list_from_url]{org_list_from_url()}},
\code{\link[=select_license]{select_license()}},
\code{\link[=store_organisations]{store_organisations()}},
\code{\link[=stored_organisations]{stored_organisations()}}
}
\concept{organisation}
