% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_individual.R
\name{select_individual}
\alias{select_individual}
\title{Select an individual from the local database or add a new individual.}
\usage{
select_individual(email, lang, org = org_list$new()$read())
}
\arguments{
\item{email}{An optional email address.
When given and it matches with a single person, the function immediately
returns the information of that person.}

\item{lang}{The language to use for the affiliation.
Defaults to the first language in the \code{name} vector of the
\code{org_list} object.
When the affiliation is not available in that language,
it will use the first available language.}

\item{org}{An \code{org_list} object containing the organisation information.
Defaults to the \code{org_list} object read from the current working directory.}
}
\value{
A data.frame with individual information.
}
\description{
Reuse existing individual information or add a new individual.
Allows to update existing individual information.
}
\examples{
\dontrun{
select_individual(lang = "en-GB")
}
}
\seealso{
Other individual:
\code{\link[=add_individual]{add_individual()}},
\code{\link[=has_person_role]{has_person_role()}},
\code{\link[=individual2badge]{individual2badge()}},
\code{\link[=individual2df]{individual2df()}},
\code{\link[=individual2person]{individual2person()}},
\code{\link[=select_person_role]{select_person_role()}},
\code{\link[=store_individuals]{store_individuals()}}
}
\concept{individual}
