% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssh_http.R
\name{ssh_http}
\alias{ssh_http}
\title{Convert SSH URL to HTTP URL}
\usage{
ssh_http(url)
}
\arguments{
\item{url}{The git URL to convert.}
}
\value{
The converted HTTP URL.
}
\description{
This function converts a git SSH URL to an HTTP URL.
It also removes any \code{OAuth2} tokens from the URL.
The resulting URL is used to determine the root URL of the organisation for
retrieving the organisation list.
}
\examples{
ssh_http("git@github.com:inbo/citeme.git")
ssh_http("https://github.com/inbo/citeme")
}
\seealso{
Other git:
\code{\link[=is_repository]{is_repository()}},
\code{\link[=is_tracked_not_modified]{is_tracked_not_modified()}}
}
\concept{git}
