% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store_authors.R
\name{store_authors}
\alias{store_authors}
\title{Store author details for later usage}
\usage{
store_authors(x = ".")
}
\arguments{
\item{x}{Path to a project}
}
\description{
Store author details for later usage
}
\seealso{
Other author: 
\code{\link{author2df}()},
\code{\link{use_author}()}
}
\concept{author}
