% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store_organisations.R
\name{stored_organisations}
\alias{stored_organisations}
\title{Retrieve stored organisation information}
\usage{
stored_organisations()
}
\value{
A data.frame with the stored organisation information.
The data.frame contains the columns: \code{email}, \code{default_name}, \code{names},
\code{ror}, \code{orcid}, \code{zenodo}, \code{website}, and \code{logo}.
Returns an empty data.frame if no organisations are stored.
}
\description{
This function retrieves organisation information that was previously stored
using \code{store_organisations()}.
}
\examples{
\dontrun{
stored_organisations()
}
}
\seealso{
Other organisation:
\code{\link[=cache_org]{cache_org()}},
\code{\link[=get_available_organisations]{get_available_organisations()}},
\code{\link[=get_default_org_list]{get_default_org_list()}},
\code{\link[=inbo_org_list]{inbo_org_list()}},
\code{\link[=new_org_item]{new_org_item()}},
\code{\link[=new_org_list]{new_org_list()}},
\code{\link[=org_list_from_url]{org_list_from_url()}},
\code{\link[=organisation2df]{organisation2df()}},
\code{\link[=select_license]{select_license()}},
\code{\link[=store_organisations]{store_organisations()}}
}
\concept{organisation}
