% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_individual.R
\name{use_individual}
\alias{use_individual}
\title{Which individual to use}
\usage{
use_individual(email, lang)
}
\arguments{
\item{email}{An optional email address.
When given and it matches with a single person, the function immediately
returns the information of that person.}

\item{lang}{The language to use for the affiliation.
Defaults to the first language in the \code{name} vector of the
\code{org_list} object.
When the affiliation is not available in that language,
it will use the first available language.}
}
\value{
A data.frame with individual information.
}
\description{
Reuse existing individual information or add a new individual.
Allows to update existing individual information.
}
\seealso{
Other individual: 
\code{\link{add_individual}()},
\code{\link{individual2df}()},
\code{\link{individual2person}()},
\code{\link{store_individuals}()}
}
\concept{individual}
