% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_citation.R
\name{validate_citation}
\alias{validate_citation}
\title{Validate a citation metadata object}
\usage{
validate_citation(meta)
}
\arguments{
\item{meta}{A \code{citation_meta} object.}
}
\value{
A character vector of validation errors,
or an empty character vector if the metadata is valid.
}
\description{
Validate a citation metadata object
}
\examples{
\dontrun{
meta <- citation_meta$new(".")
validate_citation(meta)
}
}
\seealso{
Other validation:
\code{\link[=validate_email]{validate_email()}},
\code{\link[=validate_language]{validate_language()}},
\code{\link[=validate_license_list]{validate_license_list()}},
\code{\link[=validate_orcid]{validate_orcid()}},
\code{\link[=validate_ror]{validate_ror()}},
\code{\link[=validate_url]{validate_url()}}
}
\concept{validation}
