% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_email.R
\name{validate_email}
\alias{validate_email}
\title{Check if a vector contains valid email}
\usage{
validate_email(email)
}
\arguments{
\item{email}{A vector with email addresses.}
}
\value{
A logical vector.
}
\description{
It only checks the format of the text, not if the email address exists.
}
\examples{
validate_email("user@example.com")
validate_email("invalid")
validate_email(c("a@b.com", "not-an-email"))
}
\seealso{
Other validation:
\code{\link[=validate_citation]{validate_citation()}},
\code{\link[=validate_language]{validate_language()}},
\code{\link[=validate_license_list]{validate_license_list()}},
\code{\link[=validate_orcid]{validate_orcid()}},
\code{\link[=validate_ror]{validate_ror()}},
\code{\link[=validate_url]{validate_url()}}
}
\concept{validation}
