% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_language.R
\name{validate_language}
\alias{validate_language}
\title{Validate language code}
\usage{
validate_language(language)
}
\arguments{
\item{language}{A string to validate}
}
\value{
The input language code if it is valid, otherwise an error message.
}
\description{
A valid language code is a string in the format of \code{xx-YY}, where \code{xx} is a
two-letter lowercase language code and \code{YY} is a two-letter uppercase
country code.
For example, \code{en-GB} for English (United Kingdom) and \code{nl-BE} for Dutch
(Belgium).
}
\examples{
validate_language("en-GB")
validate_language("nl-BE")
}
\seealso{
Other validation:
\code{\link[=validate_citation]{validate_citation()}},
\code{\link[=validate_email]{validate_email()}},
\code{\link[=validate_license_list]{validate_license_list()}},
\code{\link[=validate_orcid]{validate_orcid()}},
\code{\link[=validate_ror]{validate_ror()}},
\code{\link[=validate_url]{validate_url()}}
}
\concept{validation}
