% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_license.R
\name{validate_license}
\alias{validate_license}
\title{Validate a license list}
\usage{
validate_license(license)
}
\arguments{
\item{license}{A list with three named character vectors: \code{package},
\code{project}, and \code{data}.
Each vector should contain unique license names as values and unique
package, project, or data names as names.}
}
\value{
Invisibly returns \code{NULL} if the license list is valid, otherwise
throws an error.
}
\description{
Validate a license list
}
\seealso{
Other validation: 
\code{\link{validate_email}()},
\code{\link{validate_orcid}()},
\code{\link{validate_ror}()},
\code{\link{validate_url}()}
}
\concept{validation}
