% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_ror.R
\name{validate_ror}
\alias{validate_ror}
\title{Validate ROR}
\usage{
validate_ror(ror)
}
\arguments{
\item{ror}{A ROR to validate.}
}
\value{
A logical value indicating whether the ROR is valid.
}
\description{
The Research Organization Registry (\href{https://ror.org/}{ROR}) is a global,
community-led registry of open persistent identifiers for research
organizations.
ROR makes it easy for anyone or any system to disambiguate institution names
and connect research organizations to researchers and research outputs.
Validate that a ROR is a string of 9 characters starting with 0, followed by
6 characters which can be a letter (except i, l, o) or a digit, and ending
with 2 digits.
}
\examples{
validate_ror("0abcde789")
validate_ror("invalid")
}
\seealso{
Other validation:
\code{\link[=validate_citation]{validate_citation()}},
\code{\link[=validate_email]{validate_email()}},
\code{\link[=validate_language]{validate_language()}},
\code{\link[=validate_license_list]{validate_license_list()}},
\code{\link[=validate_orcid]{validate_orcid()}},
\code{\link[=validate_url]{validate_url()}}
}
\concept{validation}
