% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_badges.R
\name{add_badges}
\alias{add_badges}
\title{add badges to a README}
\usage{
add_badges(readme_path = ".", ...)
}
\arguments{
\item{readme_path}{Directory containing the \code{README.md} or \code{README.Rmd} file.}

\item{...}{Additional arguments}
}
\description{
\itemize{
\item \code{doi}: add a DOI badge
\item \code{language}: add a language badge
\item \code{license}: add a license badge
\item \code{url}: add a website badge
\item \code{version}: add a version badge
}
}
\examples{
\dontrun{
  add_badges(url = "https://www.inbo.be")
  add_badges(doi = "10.5281/zenodo.8063503")
  add_badges(version = "v0.1.2")
  add_badges(url = "https://www.inbo.be", doi = "10.5281/zenodo.8063503")
}
}
\seealso{
Other utils: 
\code{\link{coalesce}()}
}
\concept{utils}
