% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_individual.R
\name{add_individual}
\alias{add_individual}
\title{Add an individual to a file}
\usage{
add_individual(
  path = ".",
  role = c("aut", "cre", "ctb", "rev", "cph", "fnd", "pbl")
)
}
\arguments{
\item{path}{A path to a file or directory.
Supported file types:
\itemize{
\item Quarto files (\verb{_quarto.yml} or \verb{*.qmd})
\item R package DESCRIPTION files
\item README files (\code{README.md} or \code{README.Rmd})
\item Bookdown index files (\code{index.md} or \code{index.Rmd})
}}

\item{role}{The role of the person to add.
One or muliplte of \code{"aut"} (author), \code{"cre"} (creator/maintainer), \code{"ctb"}
(contributor), \code{"rev"} (reviewer), \code{"cph"} (copyright holder),
\code{"fnd"} (funder), or \code{"pbl"} (publisher).}
}
\description{
Add an individual to a file
}
\seealso{
Other individual: 
\code{\link{individual2badge}()},
\code{\link{individual2df}()},
\code{\link{individual2person}()},
\code{\link{select_individual}()},
\code{\link{store_individuals}()}
}
\concept{individual}
