% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_individual.R
\name{add_individual}
\alias{add_individual}
\title{Add an individual to a file}
\usage{
add_individual(path = ".", role = c("aut", "rev", "cph", "fnd"))
}
\arguments{
\item{path}{A path to a file or directory.
The file must be a quarto file (e.g. \verb{_quarto.yml} or \verb{*.qmd}).}

\item{role}{The role of the person to add.
One of \code{"aut"} (author), \code{"rev"} (reviewer), \code{"cph"} (copyright holder), or
\code{"fnd"} (funder).}
}
\description{
Add an individual to a file
}
\seealso{
Other individual: 
\code{\link{individual2df}()},
\code{\link{store_individuals}()},
\code{\link{use_individual}()}
}
\concept{individual}
