% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ask_email.R
\name{ask_email}
\alias{ask_email}
\title{Ask for an e-mail address}
\usage{
ask_email(prompt)
}
\arguments{
\item{prompt}{the string printed when prompting the user for input.
  Should usually end with a space \code{" "}.}
}
\description{
Ask for an e-mail address
}
\seealso{
Other question:
\code{\link[=ask_keywords]{ask_keywords()}},
\code{\link[=ask_language]{ask_language()}},
\code{\link[=ask_new_license]{ask_new_license()}},
\code{\link[=ask_orcid]{ask_orcid()}},
\code{\link[=ask_ror]{ask_ror()}},
\code{\link[=ask_url]{ask_url()}},
\code{\link[=ask_yes_no]{ask_yes_no()}},
\code{\link[=menu_first]{menu_first()}}
}
\concept{question}
