% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ask_language.R
\name{ask_language}
\alias{ask_language}
\title{Ask for a language}
\usage{
ask_language(org, prompt = "Which language?")
}
\arguments{
\item{org}{An \code{org_list} object containing the available languages.}

\item{prompt}{the string printed when prompting the user for input.
  Should usually end with a space \code{" "}.}
}
\description{
Ask for a language
}
\seealso{
Other question: 
\code{\link{ask_email}()}
}
\concept{question}
