% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ask_new_license.R
\name{ask_new_license}
\alias{ask_new_license}
\title{Ask one or more licenses.}
\usage{
ask_new_license(licenses, type = c("package", "project", "data"))
}
\arguments{
\item{licenses}{A named vector of available licenses.
The vector contains the URL to the markdown version of the license.
Use the abbreviation of the license name as names.}

\item{type}{The type of license.
Must be one of \code{"package"}, \code{"project"} or \code{"data"}.}
}
\value{
A named character vector where each element is a URL to the
markdown version of the license and each name is the license abbreviation.
Returns an empty character vector when no license is selected.
}
\description{
Ask one or more licenses.
}
\examples{
\dontrun{
ask_new_license(
  c("MIT" = "https://example.com/mit.md"),
  type = "package"
)
}
}
\seealso{
Other question:
\code{\link[=ask_email]{ask_email()}},
\code{\link[=ask_keywords]{ask_keywords()}},
\code{\link[=ask_language]{ask_language()}},
\code{\link[=ask_orcid]{ask_orcid()}},
\code{\link[=ask_ror]{ask_ror()}},
\code{\link[=ask_url]{ask_url()}},
\code{\link[=ask_yes_no]{ask_yes_no()}},
\code{\link[=menu_first]{menu_first()}}
}
\concept{question}
