% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ask_new_license.R
\name{ask_new_license}
\alias{ask_new_license}
\title{Ask one or more licenses.}
\usage{
ask_new_license(licenses, type = c("package", "project", "data"))
}
\arguments{
\item{licenses}{A named vector of available licenses.
The vector contains the URL to the markdown version of the license.
Use the abbreviation of the license name as names.}

\item{type}{The type of license.
Must be one of \code{"package"}, \code{"project"} or \code{"data"}.}
}
\description{
Ask one or more licenses.
}
\seealso{
Other question: 
\code{\link{ask_email}()},
\code{\link{ask_language}()},
\code{\link{ask_ror}()}
}
\concept{question}
