% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ask_orcid.R
\name{ask_orcid}
\alias{ask_orcid}
\title{Ask for an \verb{ORCID iD}}
\usage{
ask_orcid(prompt = "orcid: ")
}
\arguments{
\item{prompt}{A character string to display as a prompt to the user.
The default is \code{"orcid: "}.}
}
\value{
A character string containing the \verb{ORCID iD} entered by the user.
}
\description{
The \href{https://orcid.org/}{\verb{ORCID iD}} is a unique, persistent identifier
free of charge to researchers.
This function prompts the user to enter an \verb{ORCID iD} and validates its
format.
The \verb{ORCID iD} must be in the format \code{0000-0000-0000-0000} where the last
digit can be a number or \code{"X"}.
Empty strings are considered valid to allow optional \verb{ORCID iD}.
}
\seealso{
Other question: 
\code{\link{ask_email}()},
\code{\link{ask_language}()},
\code{\link{ask_new_license}()},
\code{\link{ask_ror}()},
\code{\link{ask_url}()},
\code{\link{ask_yes_no}()},
\code{\link{menu_first}()}
}
\concept{question}
