% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ask_url.R
\name{ask_url}
\alias{ask_url}
\title{Ask an URL}
\usage{
ask_url(prompt)
}
\arguments{
\item{prompt}{the string printed when prompting the user for input.
  Should usually end with a space \code{" "}.}
}
\description{
Ask an URL
}
\seealso{
Other question: 
\code{\link{ask_email}()},
\code{\link{ask_language}()},
\code{\link{ask_new_license}()},
\code{\link{ask_ror}()}
}
\concept{question}
