% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ask_yes_no.R
\name{ask_yes_no}
\alias{ask_yes_no}
\title{Function to ask a simple yes no question}
\usage{
ask_yes_no(msg, default = TRUE, prompts = c("Yes", "No", "Cancel"), ...)
}
\arguments{
\item{msg}{
The prompt message for the user.
}

\item{default}{
The default response.
}

\item{prompts}{
Any of:  a character vector containing 3 prompts corresponding to
return values of \code{TRUE}, \code{FALSE}, or \code{NA}, or
a single character value containing the prompts separated by
\code{/} characters, or a function to call.
}

\item{...}{
Additional parameters, ignored by the default function.
}
}
\value{
A logical value: \code{TRUE} for yes, \code{FALSE} for no, or \code{NA} for
cancel.
Returns \code{default} in non-interactive sessions.
}
\description{
Provides a simple wrapper around \code{utils::askYesNo()}.
This function is used to ask yes no questions in an interactive way.
It repeats the question until a valid answer is given.
}
\examples{
# Returns default in non-interactive sessions
ask_yes_no("Continue?", default = TRUE)
}
\seealso{
Other question:
\code{\link[=ask_email]{ask_email()}},
\code{\link[=ask_keywords]{ask_keywords()}},
\code{\link[=ask_language]{ask_language()}},
\code{\link[=ask_new_license]{ask_new_license()}},
\code{\link[=ask_orcid]{ask_orcid()}},
\code{\link[=ask_ror]{ask_ror()}},
\code{\link[=ask_url]{ask_url()}},
\code{\link[=menu_first]{menu_first()}}
}
\concept{question}
