% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_available_organisations.R
\name{get_available_organisations}
\alias{get_available_organisations}
\title{Get the list of available organisations}
\usage{
get_available_organisations()
}
\value{
A list containing the names, languages, licenses, ORCID, Zenodo, ROR,
website, and logo of the available organisations.
}
\description{
This function retrieves the list of available organisations from the local
configuration files.
It returns a list containing the names, languages, licenses, ORCID, Zenodo,
ROR, website, and logo of the available organisations.
}
\seealso{
Other organisation: 
\code{\link{cache_org}()},
\code{\link{get_default_org_list}()},
\code{\link{inbo_org_list}()},
\code{\link{new_org_list}()},
\code{\link{org_list_from_url}()},
\code{\link{select_license}()}
}
\concept{organisation}
