% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_default_org_list.R
\name{get_default_org_list}
\alias{get_default_org_list}
\title{Get the default organization list}
\usage{
get_default_org_list(x = ".")
}
\arguments{
\item{x}{The path to the repository.
Defaults to the current working directory.}
}
\value{
An \code{org_list} object containing the organisation list.
The function also stores the information in the user's R configuration.
}
\description{
This function retrieves the default organisation list from the
\code{organisation.yml} file in the organisations \code{checklist} repository.
The \code{origin} of the repository is used to determine the root URL of the
organisation.
}
\seealso{
Other organisation: 
\code{\link{inbo_org_list}()},
\code{\link{new_org_list}()}
}
\concept{organisation}
