% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/org_list_class.R
\name{inbo_org_list}
\alias{inbo_org_list}
\title{The INBO organisation list}
\usage{
inbo_org_list()
}
\description{
The INBO organisation list
}
\seealso{
Other organisation: 
\code{\link{get_default_org_list}()},
\code{\link{new_org_list}()},
\code{\link{org_list_from_url}()}
}
\concept{organisation}
