% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individual2badge.R
\name{individual2badge}
\alias{individual2badge}
\title{Create a markdown badge for an individual}
\usage{
individual2badge(
  individual,
  role = c("aut", "cre", "cph", "ctb", "fnd", "rev")
)
}
\arguments{
\item{individual}{A data frame with columns \code{given}, \code{family}, \code{orcid},
\code{email}, and \code{affiliation}.}

\item{role}{A character string indicating the individual's role.
Must be one or more of \code{"aut"} (author), \code{"cre"} (contact person),
\code{"cph"} (copyright holder), \code{"ctb"} (contributor), \code{"fnd"} (funder),
\code{"rev"} (reviewer).
Default is `"aut"``.}
}
\value{
A character string containing the markdown badge for the individual.
}
\description{
This function creates a markdown badge for an individual, including their
name, ORCID, email, and affiliation if available.
It also adds a footnote with the individual's role and affiliation.
}
\seealso{
Other individual: 
\code{\link{add_individual}()},
\code{\link{individual2df}()},
\code{\link{individual2person}()},
\code{\link{store_individuals}()},
\code{\link{use_individual}()}
}
\concept{individual}
