% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store_individuals.R
\name{individual2df}
\alias{individual2df}
\title{Convert person object in a data.frame.}
\usage{
individual2df(person)
}
\arguments{
\item{person}{The person object or a list of person objects, \code{NA} or \code{NULL}.
Any \code{"character"} is converted to a person object using \code{as.person()} with a
warning.}
}
\value{
A \code{data.frame} with columns \code{given}, \code{family}, \code{email}, \code{orcid},
\code{ror}, \code{affiliation}, and \code{role}.
}
\description{
Results in a data.frame with the given name, family name, e-mail, ORCID,
affiliation and role.
Missing elements result in an empty string (\code{""}).
Persons with multiple roles will have the roles as a comma separated string.
}
\examples{
individual2df(person("Jane", "Doe", role = "aut"))
individual2df(NULL)
}
\seealso{
Other individual:
\code{\link[=add_individual]{add_individual()}},
\code{\link[=has_person_role]{has_person_role()}},
\code{\link[=individual2badge]{individual2badge()}},
\code{\link[=individual2person]{individual2person()}},
\code{\link[=select_individual]{select_individual()}},
\code{\link[=select_person_role]{select_person_role()}},
\code{\link[=store_individuals]{store_individuals()}}
}
\concept{individual}
