% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store_individuals.R
\name{individual2df}
\alias{individual2df}
\title{Convert person object in a data.frame.}
\usage{
individual2df(person)
}
\arguments{
\item{person}{The person object or a list of person objects, \code{NA} or \code{NULL}.
Any \code{"character"} is converted to a person object using \code{as.person()} with a
warning.}
}
\description{
Results in a data.frame with the given name, family name, e-mail, ORCID,
affiliation and role.
Missing elements result in an empty string (\code{""}).
Persons with multiple roles will have the roles as a comma separated string.
}
\seealso{
Other individual: 
\code{\link{add_individual}()},
\code{\link{individual2badge}()},
\code{\link{individual2person}()},
\code{\link{store_individuals}()},
\code{\link{use_individual}()}
}
\concept{individual}
