% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individual2person.R
\name{individual2person}
\alias{individual2person}
\title{Select the individual information to use as a \code{person} object}
\usage{
individual2person(role = "aut", lang)
}
\arguments{
\item{role}{The role to use for the \code{person} object.
Defaults to `"aut"`` (author).}

\item{lang}{The language to use for the affiliation.}
}
\description{
This function retrieves the individual information using \code{use_individual()}
and converts it to a \code{person} object.
}
\seealso{
Other individual: 
\code{\link{add_individual}()},
\code{\link{individual2df}()},
\code{\link{store_individuals}()},
\code{\link{use_individual}()}
}
\concept{individual}
