% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_tracked_not_modified.R
\name{is_tracked_not_modified}
\alias{is_tracked_not_modified}
\title{Check if a file is tracked and not modified}
\usage{
is_tracked_not_modified(file, repo = ".")
}
\arguments{
\item{file}{path relative to the git root directory.}

\item{repo}{path to the repository}
}
\value{
A logical value: \code{TRUE} when the file is tracked by git and has no
modifications, \code{FALSE} otherwise.
Returns \code{TRUE} when no git repository is found.
}
\description{
Check if a file is tracked and not modified
}
\seealso{
Other git:
\code{\link[=is_repository]{is_repository()}},
\code{\link[=ssh_http]{ssh_http()}}
}
\concept{git}
