% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/menu_first.R
\name{menu_first}
\alias{menu_first}
\title{Improved version of \code{utils::menu()}}
\usage{
menu_first(choices, graphics = FALSE, title = NULL)
}
\arguments{
\item{choices}{a character vector of choices}

\item{graphics}{a logical indicating whether a graphics menu should be
    used if available.}

\item{title}{a character string to be used as the title of the menu.
    \code{NULL} is also accepted.}
}
\value{
An integer indicating the index of the selected choice.
Returns \code{1} in non-interactive sessions.
}
\description{
This function is a wrapper around \code{utils::menu()} that returns the index of
the first choice instead of the index of the selected choice.
This is useful when you want to ask a question with only one choice and you
want to return a logical value.
}
\examples{
# Returns 1 in non-interactive sessions
menu_first(choices = c("option A", "option B"))
}
\seealso{
Other question:
\code{\link[=ask_email]{ask_email()}},
\code{\link[=ask_keywords]{ask_keywords()}},
\code{\link[=ask_language]{ask_language()}},
\code{\link[=ask_new_license]{ask_new_license()}},
\code{\link[=ask_orcid]{ask_orcid()}},
\code{\link[=ask_ror]{ask_ror()}},
\code{\link[=ask_url]{ask_url()}},
\code{\link[=ask_yes_no]{ask_yes_no()}}
}
\concept{question}
