% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/menu_first.R, R/use_author.R
\name{menu_first}
\alias{menu_first}
\title{Improved version of menu()
This function is a wrapper around \code{utils::menu()} that returns the index of
the first choice instead of the index of the selected choice.
This is useful when you want to ask a question with only one choice and you
want to return a logical value.}
\usage{
menu_first(choices, graphics = FALSE, title = NULL)

menu_first(choices, graphics = FALSE, title = NULL)
}
\arguments{
\item{choices}{a character vector of choices}

\item{graphics}{a logical indicating whether a graphics menu should be
    used if available.}

\item{title}{a character string to be used as the title of the menu.
    \code{NULL} is also accepted.}
}
\description{
Improved version of menu()
This function is a wrapper around \code{utils::menu()} that returns the index of
the first choice instead of the index of the selected choice.
This is useful when you want to ask a question with only one choice and you
want to return a logical value.

Improved version of menu()
}
\seealso{
Other question: 
\code{\link{ask_email}()},
\code{\link{ask_language}()},
\code{\link{ask_new_license}()},
\code{\link{ask_orcid}()},
\code{\link{ask_ror}()},
\code{\link{ask_url}()},
\code{\link{ask_yes_no}()}

Other utils: 
\code{\link{author2df}()},
\code{\link{get_default_org_list}()},
\code{\link{inbo_org_list}()},
\code{\link{store_authors}()},
\code{\link{use_author}()},
\code{\link{validate_orcid}()}
}
\concept{question}
\concept{utils}
