% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_org_item.R
\name{new_org_item}
\alias{new_org_item}
\title{Interactively create a new \code{org_item}}
\usage{
new_org_item(languages, licenses)
}
\arguments{
\item{languages}{A character vector of language codes.}

\item{licenses}{A list of license items.}
}
\value{
An \code{org_item} object.
}
\description{
Interactively create a new \code{org_item}
}
\seealso{
Other organisation:
\code{\link[=cache_org]{cache_org()}},
\code{\link[=get_available_organisations]{get_available_organisations()}},
\code{\link[=get_default_org_list]{get_default_org_list()}},
\code{\link[=inbo_org_list]{inbo_org_list()}},
\code{\link[=new_org_list]{new_org_list()}},
\code{\link[=org_list_from_url]{org_list_from_url()}},
\code{\link[=organisation2df]{organisation2df()}},
\code{\link[=select_license]{select_license()}},
\code{\link[=store_organisations]{store_organisations()}},
\code{\link[=stored_organisations]{stored_organisations()}}
}
\concept{organisation}
