% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_org_item.R
\name{new_org_item}
\alias{new_org_item}
\title{Create a new org item}
\usage{
new_org_item(languages, licenses)
}
\arguments{
\item{languages}{A character vector of language codes.}

\item{licenses}{A list of license items.}
}
\value{
An org item.
}
\description{
Create a new org item
}
\seealso{
Other organisation: 
\code{\link{cache_org}()},
\code{\link{get_available_organisations}()},
\code{\link{get_default_org_list}()},
\code{\link{inbo_org_list}()},
\code{\link{new_org_list}()},
\code{\link{org_list_from_url}()},
\code{\link{select_license}()}
}
\concept{organisation}
