% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_org_list.R
\name{new_org_list}
\alias{new_org_list}
\title{Interactively create a new organisation list}
\usage{
new_org_list(git)
}
\arguments{
\item{git}{An optional string with the absolute path to a git
organisation.
E.g. \code{"https://github.com/inbo"}}
}
\description{
An interactive alternative for \code{org_list$new()}.
Reuses available organisations where possible.
}
\seealso{
\code{\link{org_list}}, \code{\link{org_item}}

Other organisation: 
\code{\link{cache_org}()},
\code{\link{get_available_organisations}()},
\code{\link{get_default_org_list}()},
\code{\link{inbo_org_list}()},
\code{\link{new_org_item}()},
\code{\link{org_list_from_url}()},
\code{\link{select_license}()}
}
\concept{organisation}
