% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/org_list_from_url.R
\name{org_list_from_url}
\alias{org_list_from_url}
\title{Get the default organisation list from a git URL
This function retrieves the default organisation list from a git URL.
It checks if the organisation list is already cached in the user's R
configuration.
If it is, it returns the cached version.
If not, it attempts to retrieve the organisation list from the specified git
URL and caches it for future use.}
\usage{
org_list_from_url(git)
}
\arguments{
\item{git}{The git URL to retrieve the organisation list from.}
}
\value{
An \code{org_list} object containing the organisation list.
}
\description{
Get the default organisation list from a git URL
This function retrieves the default organisation list from a git URL.
It checks if the organisation list is already cached in the user's R
configuration.
If it is, it returns the cached version.
If not, it attempts to retrieve the organisation list from the specified git
URL and caches it for future use.
}
\seealso{
Other organisation: 
\code{\link{cache_org}()},
\code{\link{get_default_org_list}()},
\code{\link{inbo_org_list}()},
\code{\link{new_org_list}()}
}
\concept{organisation}
