% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_license.R
\name{select_license}
\alias{select_license}
\title{Select a license for a project, package or dataset}
\usage{
select_license(org, type = c("package", "project", "data"))
}
\arguments{
\item{org}{An object of class \code{org_list} created with \code{new_org_list()} or
\code{org_list$new()}.}

\item{type}{The type of license to select. One of \code{"package"}, \code{"project"},
or \code{"data"}.
Default is \code{"package"}.}
}
\value{
The name of the selected license.
}
\description{
This function allows you to select a license for a project, package or
dataset from the list of allowed licenses for the organisation.
If there is only one license available, it will be selected automatically.
}
\seealso{
Other organisation: 
\code{\link{cache_org}()},
\code{\link{get_available_organisations}()},
\code{\link{get_default_org_list}()},
\code{\link{inbo_org_list}()},
\code{\link{new_org_item}()},
\code{\link{new_org_list}()},
\code{\link{org_list_from_url}()}
}
\concept{organisation}
