% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_person_role.R
\name{select_person_role}
\alias{select_person_role}
\title{Select the person object with a given role}
\usage{
select_person_role(individual, role)
}
\arguments{
\item{individual}{A \code{person} object.}

\item{role}{A character vector of roles to check for.}
}
\value{
A logical vector indicating whether the individual has any of the
specified roles.
}
\description{
Select the person object with a given role
}
\seealso{
Other individual:
\code{\link[=add_individual]{add_individual()}},
\code{\link[=has_person_role]{has_person_role()}},
\code{\link[=individual2badge]{individual2badge()}},
\code{\link[=individual2df]{individual2df()}},
\code{\link[=individual2person]{individual2person()}},
\code{\link[=select_individual]{select_individual()}},
\code{\link[=store_individuals]{store_individuals()}}
}
\concept{individual}
