% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store_authors.R
\name{store_authors}
\alias{store_authors}
\title{Store author details for later usage}
\usage{
store_authors(x = ".")
}
\arguments{
\item{x}{Path to a project}
}
\description{
Store author details for later usage
}
\seealso{
Other utils: 
\code{\link{author2df}()},
\code{\link{get_default_org_list}()},
\code{\link{inbo_org_list}()},
\code{\link{menu_first}()},
\code{\link{use_author}()},
\code{\link{validate_orcid}()}
}
\concept{utils}
