% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store_individuals.R
\name{store_individuals}
\alias{store_individuals}
\title{Store individual details for later usage}
\usage{
store_individuals(x = ".")
}
\arguments{
\item{x}{Path to a project}
}
\value{
Invisibly returns \code{NULL}.
Writes the individual details to the user data directory as a side effect.
}
\description{
Store individual details for later usage
}
\seealso{
Other individual:
\code{\link[=add_individual]{add_individual()}},
\code{\link[=has_person_role]{has_person_role()}},
\code{\link[=individual2badge]{individual2badge()}},
\code{\link[=individual2df]{individual2df()}},
\code{\link[=individual2person]{individual2person()}},
\code{\link[=select_individual]{select_individual()}},
\code{\link[=select_person_role]{select_person_role()}}
}
\concept{individual}
