% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store_individuals.R
\name{store_individuals}
\alias{store_individuals}
\title{Store individual details for later usage}
\usage{
store_individuals(x = ".")
}
\arguments{
\item{x}{Path to a project}
}
\description{
Store individual details for later usage
}
\seealso{
Other individual: 
\code{\link{individual2df}()},
\code{\link{use_individual}()}
}
\concept{individual}
