% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store_organisations.R
\name{store_organisations}
\alias{store_organisations}
\title{Store organisation details for later usage}
\usage{
store_organisations(x)
}
\arguments{
\item{x}{An \code{org_list} or \code{org_item} object to store.}
}
\value{
\code{NULL} invisibly.
}
\description{
This function stores organisation information locally for later retrieval.
The organisations are stored in a tab-separated file in the user's data
directory.
}
\seealso{
Other organisation:
\code{\link[=cache_org]{cache_org()}},
\code{\link[=get_available_organisations]{get_available_organisations()}},
\code{\link[=get_default_org_list]{get_default_org_list()}},
\code{\link[=inbo_org_list]{inbo_org_list()}},
\code{\link[=new_org_item]{new_org_item()}},
\code{\link[=new_org_list]{new_org_list()}},
\code{\link[=org_list_from_url]{org_list_from_url()}},
\code{\link[=organisation2df]{organisation2df()}},
\code{\link[=select_license]{select_license()}},
\code{\link[=stored_organisations]{stored_organisations()}}
}
\concept{organisation}
