% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_citation.R
\name{validate_citation}
\alias{validate_citation}
\title{Validate a citation metadata object}
\usage{
validate_citation(meta)
}
\arguments{
\item{meta}{A \code{citation_meta} object.}
}
\value{
A character vector of validation errors,
or an empty character vector if the metadata is valid.
}
\description{
Validate a citation metadata object
}
