% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_email.R
\name{validate_email}
\alias{validate_email}
\title{Check if a vector contains valid email}
\usage{
validate_email(email)
}
\arguments{
\item{email}{A vector with email addresses.}
}
\value{
A logical vector.
}
\description{
It only checks the format of the text, not if the email address exists.
}
\seealso{
Other validation: 
\code{\link{validate_license}()},
\code{\link{validate_orcid}()},
\code{\link{validate_ror}()},
\code{\link{validate_url}()}
}
\concept{validation}
