% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_email}
\alias{validate_email}
\title{Check if a vector contains valid email}
\usage{
validate_email(email)
}
\arguments{
\item{email}{A vector with email addresses.}
}
\value{
A logical vector.
}
\description{
It only checks the format of the text, not if the email address exists.
}
\seealso{
Other utils: 
\code{\link{author2df}()},
\code{\link{get_default_org_list}()},
\code{\link{inbo_org_list}()},
\code{\link{menu_first}()},
\code{\link{store_authors}()},
\code{\link{use_author}()},
\code{\link{validate_orcid}()}
}
\concept{utils}
