% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_author.R, R/utils.R
\name{validate_orcid}
\alias{validate_orcid}
\title{Validate the structure of an ORCID id}
\usage{
validate_orcid(orcid)

validate_orcid(orcid)
}
\arguments{
\item{orcid}{A vector of ORCID}
}
\value{
A logical vector with the same length as the input vector.

A logical vector with the same length as the input vector.
}
\description{
Checks whether the ORCID has the proper format and the checksum.

Checks whether the ORCID has the proper format and the checksum.
}
\seealso{
Other utils: 
\code{\link{author2df}()},
\code{\link{get_default_org_list}()},
\code{\link{inbo_org_list}()},
\code{\link{menu_first}()},
\code{\link{store_authors}()},
\code{\link{use_author}()},
\code{\link{validate_email}()}

Other utils: 
\code{\link{author2df}()},
\code{\link{get_default_org_list}()},
\code{\link{inbo_org_list}()},
\code{\link{menu_first}()},
\code{\link{store_authors}()},
\code{\link{use_author}()},
\code{\link{validate_email}()}
}
\concept{utils}
