% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_orcid.R
\name{validate_orcid}
\alias{validate_orcid}
\title{Validate the structure of an \verb{ORCID iD}}
\usage{
validate_orcid(orcid)
}
\arguments{
\item{orcid}{A vector of ORCID}
}
\value{
A logical vector with the same length as the input vector.
}
\description{
The \href{https://orcid.org/}{\verb{ORCID iD}} is a unique, persistent identifier
free of charge to researchers.
Checks whether the \verb{ORCID iD} has the proper format and a correct checksum.
The \verb{ORCID iD} must be in the format \code{0000-0000-0000-0000} where the last
digit can be a number or \code{"X"}.
Empty strings are considered valid to allow optional \verb{ORCID iD}.
}
\seealso{
Other validation: 
\code{\link{validate_email}()},
\code{\link{validate_license}()},
\code{\link{validate_ror}()},
\code{\link{validate_url}()}
}
\concept{validation}
