% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_author.R, R/validate_orcid.R
\name{validate_orcid}
\alias{validate_orcid}
\title{Validate the structure of an ORCID id}
\usage{
validate_orcid(orcid)

validate_orcid(orcid)
}
\arguments{
\item{orcid}{A vector of ORCID}
}
\value{
A logical vector with the same length as the input vector.

A logical vector with the same length as the input vector.
}
\description{
Checks whether the ORCID has the proper format and the checksum.

The \href{https://orcid.org/}{\verb{ORCID iD}} is a unique, persistent identifier
free of charge to researchers.
Checks whether the \verb{ORCID iD} has the proper format and a correct checksum.
The \verb{ORCID iD} must be in the format \code{0000-0000-0000-0000} where the last
digit can be a number or \code{"X"}.
Empty strings are considered valid to allow optional \verb{ORCID iD}.
}
\seealso{
Other utils: 
\code{\link{author2df}()},
\code{\link{get_default_org_list}()},
\code{\link{inbo_org_list}()},
\code{\link{menu_first}()},
\code{\link{store_authors}()},
\code{\link{use_author}()}

Other validation: 
\code{\link{validate_email}()},
\code{\link{validate_license}()},
\code{\link{validate_ror}()},
\code{\link{validate_url}()}
}
\concept{utils}
\concept{validation}
