% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_url.R
\name{validate_url}
\alias{validate_url}
\title{Validate URL}
\usage{
validate_url(url)
}
\arguments{
\item{url}{A URL to validate.}
}
\value{
A logical value indicating whether the URL is valid.
}
\description{
Validate that a URL is a string of length 1 which is not NA and matches the
pattern of a valid URL.
}
\seealso{
Other validation:
\code{\link[=validate_citation]{validate_citation()}},
\code{\link[=validate_email]{validate_email()}},
\code{\link[=validate_language]{validate_language()}},
\code{\link[=validate_license_list]{validate_license_list()}},
\code{\link[=validate_orcid]{validate_orcid()}},
\code{\link[=validate_ror]{validate_ror()}}
}
\concept{validation}
