% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_url.R
\name{validate_url}
\alias{validate_url}
\title{Validate URL
Validate that a URL is a string of length 1 which is not NA and matches the
pattern of a valid URL.}
\usage{
validate_url(url)
}
\arguments{
\item{url}{A URL to validate.}
}
\value{
A logical value indicating whether the URL is valid.
}
\description{
Validate URL
Validate that a URL is a string of length 1 which is not NA and matches the
pattern of a valid URL.
}
\seealso{
Other validation: 
\code{\link{validate_ror}()}
}
\concept{validation}
